+++ src/test/java/de/bild/codec/dividedclassmodel/basemodel/BaseInterface.java
@@ -1,0 +1,4 @@
+package de.bild.codec.dividedclassmodel.basemodel;
+
+public interface BaseInterface {
+}