+++ src/test/java/de/bild/codec/dividedclassmodel/nonmodel/NonModelInterface.java
@@ -1,0 +1,4 @@
+package de.bild.codec.dividedclassmodel.nonmodel;
+
+public interface NonModelInterface {
+}