+++ src/test/java/de/bild/codec/dividedclassmodel/submodel/SubPojoWithinModel.java
@@ -1,0 +1,6 @@
+package de.bild.codec.dividedclassmodel.submodel;
+
+import de.bild.codec.dividedclassmodel.nonmodel.NonModelPojoExt;
+
+public class SubPojoWithinModel extends NonModelPojoExt {
+}